Moved notifications functions to own file, has broken things
